Modified the simulation to use doubles for time, removed the tick event in favor of the car generation event.

+++ src/main/java/com/smartcity/event/Event.java
@@ -5,17 +5,15 @@
  */
 package com.smartcity.event;
 
-import java.math.BigInteger;
-
 /**
  *
  * @author Blake
  */
 public abstract class Event {
     
-    public final BigInteger eventTime;
+    public final double eventTime;
     
-    public Event(BigInteger eventTime){
+    public Event(double eventTime){
         this.eventTime = eventTime;
     }
